requirements.txt
alembic==1.13.1
alembic-postgresql-enum==1.2.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
tzlocal==5.2
uvicorn==0.29.0
Werkzeug==3.0.2
